Update the feature to  display comment on submit
